refactor alot of code
use FutureBuilder to wait for data from APIs, and handle exceptions
better.
add exceptions dir and all files inside to handle exceptions.
Note: there is some unhandled exception printed in the terminal but does
not crash the app, and does not seem to be catchable as the source is
the files from the http module